Add color definition by log record attribute
Now it is possible to define coloring for each attribute of a LogRecord
individually.

There is some backwards compatability for the __setitem__ method, which
still updates the color definition of the 'levelname' attribute.

With the new set(level, **attrs) method, there can be one color definition
for each log record attribute and for each log level.